Rename "fixpoint_sdk" to "fixpoint"
