+++ pyproject.toml
@@ -1,13 +1,6 @@
 [project]
-name = "fixpoint_sdk"
+name = "fixpoint"
 version = "0.9.0"
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[tool.hatch.metadata]
-packages = [{include = "fixpoint"}, {include = "fixpoint_sdk"}]
 
 authors = [
   {name="Fixpoint", email="team@fixpoint.co"},
@@ -52,3 +45,7 @@ dev = [
 [project.urls]
 Homepage = "https://github.com/gofixpoint/python-sdk"
 Issues = "https://github.com/gofixpoint/python-sdk/issues"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
