+++ src/fixpoint/_mock_completions.py
