+++ src/fixpoint/completions.py
