+++ src/fixpoint/lib/iterwrapper.py
