+++ src/fixpoint/openapi/exceptions.py
